requirements.txt
jupyter
jupyterthemes
numpy
seaborn
svg.path

nbstripout